requirements.txt
streamlit
fpdf2


